// agricultural-pesticide-platform/types/entities.ts
export interface User {
  user_id: string
  username: string
  email: string
  full_name: string
  phone_number: string
  avatar?: string
  cccd?: string
  is_active: boolean
  created_at: Date
  updated_at: Date
  role: Role
}
// ...
export interface Role {
  role_id: string
  role_name: "ADMIN" | "DISTRIBUTOR" | "CUSTOMER"
  description: string
  is_active: boolean
}

export interface Product {
  product_id: string
  distributor: User
  categories: Category[]
  product_name: string
  description: string
  usage_instructions: string
  unit_product_price: number
  is_active: boolean
  created_at: Date
  updated_at: Date
  product_images?: ProductImage[]
  product_ingredients?: ProductIngredient[]
  product_diseases?: ProductDisease[]
}
// ...
export interface BatchProduct {
  batch_id: string
  product: Product
  batch_number: string
  quantity: number
  manufactured_date: Date
  expiry_date: Date
  low_stock_threshold: number
  is_active: boolean
}
// ...
export interface Invenstory {
  invenstory_id: string
  distributor: User // The distributor (user) who owns the inventory
  name: string // Store name
  business_license: string // Business license number
  invenstory_address: string // Full address
  invenstory_lat?: number // Latitude
  invenstory_lng?: number // Longitude
  invenstory_img?: string // Store image (URL)
  is_locked?: boolean // Warehouse lock status
  created_at: Date
  updated_at: Date
  batch_products?: BatchProduct[] // Optional list of batch products in the inventory
}